+++ lib/kubernetes-deploy/errors.rb
@@ -23,6 +23,14 @@ module KubernetesDeploy
 
   class DeploymentTimeoutError < FatalDeploymentError; end
 
+  class EjsonPrunableError < FatalDeploymentError
+    def initialize
+      super("Found #{KubernetesResource::LAST_APPLIED_ANNOTATION} annotation on " \
+          "#{EjsonSecretProvisioner::EJSON_KEYS_SECRET} secret. " \
+          "kubernetes-deploy will not continue since it is extremely unlikely that this secret should be pruned.")
+    end
+  end
+
   module Errors
     extend self
     def server_version_warning(server_version)